Add unit tests for storage initializer
