+++ storage/index.js
@@ -14,69 +14,7 @@
 
 'use strict';
 
-const path = require('path');
-const {
-	BaseEntity,
-	Account,
-	Delegate,
-	Block,
-	Transaction,
-} = require('./entities');
-const PgpAdapter = require('./adapters/pgp_adapter');
-
-class Storage {
-	constructor(options, logger) {
-		this.options = options;
-		this.logger = logger;
-
-		if (typeof Storage.instance === 'object') {
-			return Storage.instance;
-		}
-
-		this.isReady = false;
-
-		Storage.instance = this;
-
-		Storage.instance.BaseEntity = BaseEntity;
-		Storage.instance.BaseEntity.adapter = null;
-	}
-
-	/**
-	 * @return Promise
-	 */
-	bootstrap() {
-		const adapter = new PgpAdapter(
-			Object.assign({}, this.options, {
-				inTest: process.env.NODE_ENV === 'test',
-				sqlDirectory: path.join(path.dirname(__filename), './sql'),
-				logger: this.logger,
-			})
-		);
-
-		return adapter.connect().then(status => {
-			if (status) {
-				this.isReady = true;
-				Storage.instance.adapter = adapter;
-				BaseEntity.prototype.adapter = adapter;
-
-				Storage.instance.entities = {
-					Transaction: new Transaction(),
-					Block: new Block(),
-					Account: new Account(),
-					Delegate: new Delegate(),
-				};
-			}
-
-			return status;
-		});
-	}
-
-	cleanup() {
-		return Storage.instance.adapter.disconnect().then(() => {
-			this.isReady = false;
-		});
-	}
-}
+const Storage = require('./storage');
 
 module.exports = function createStorage(options, logger) {
 	return new Storage(options, logger);
